a bit of refactoring, minor improvements
