+++ src/main/java/com/ngs/stash/externalhooks/hook/ExternalPreReceiveHook.java
@@ -28,7 +28,7 @@ import com.atlassian.bitbucket.user.SecurityService;
 import com.atlassian.event.api.EventListener;
 import com.atlassian.sal.api.pluginsettings.PluginSettingsFactory;
 import com.atlassian.upm.api.license.PluginLicenseManager;
-import com.ngs.stash.externalhooks.ExternalHooksSettingsDao;
+import com.ngs.stash.externalhooks.dao.ExternalHooksSettingsDao;
 
 public class ExternalPreReceiveHook
     implements PreRepositoryHook<RepositoryHookRequest>, SettingsValidator {
@@ -72,6 +72,7 @@ public class ExternalPreReceiveHook
       SecurityService securityService)
       throws IOException {
     ExternalHooksSettingsDao settingsDao = new ExternalHooksSettingsDao(pluginSettingsFactory);
+
     return new ExternalHookScript(
         authenticationContext,
         permissions,
